fix: restructuring actions
